Fix regex to replace version in manifest
This regex is used to find (textually, not via toml parser) where to replace the
version number in the Manifest.

Until now it was only supporting version numbers containing dots and digits.
I changed that to version numbers containing anything but ~ and " to handle version
numbers with dashes, or manually written version numbers that might be invalid (if
someone litteraly writes `version = "x.y~ynh1"). We don’t actually care about the
version number we match against because we just replace it.